requirements.txt
flask
pyTelegramBotAPI
firebase-admin
gunicorn
python-dotenv